# lib/api-zod/src/generated/types/conversationWithUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Conversation } from "./conversation";
import type { User } from "./user";

export type ConversationWithUser = Conversation & {
  otherUser: User;
  unreadCount: number;
};