# lib/api-zod/src/generated/types/createTaskRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTaskRequest {
  title: string;
  description: string;
  category: string;
  categoryEmoji?: string;
  price: number;
  estimatedTime: string;
  location: string;
  isRemote: boolean;
  lat?: string;
  lng?: string;
  tags?: string[];
  creatorId: string;
}