# lib/api-zod/src/generated/types/rankingEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RankingEntry {
  id: string;
  name: string;
  avatarInitials: string;
  avatarBg: string;
  avatarColor: string;
  rankLevel: number;
  rankPoints: number;
  tasksCompleted: number;
  totalEarned: number;
}