# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: string;
  title: string;
  description: string;
  category: string;
  categoryEmoji: string;
  price: number;
  estimatedTime: string;
  location: string;
  isRemote: boolean;
  lat?: string | null;
  lng?: string | null;
  tags: string[];
  status: string;
  highlight: boolean;
  creatorId: string;
  executorId?: string | null;
  createdAt: Date;
  updatedAt: Date;
}