# lib/api-zod/src/generated/types/taskWithCreator.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from "./task";
import type { User } from "./user";

export type TaskWithCreator = Task & {
  creator: User;
};