# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  avatarInitials: string;
  avatarBg: string;
  avatarColor: string;
  bio?: string | null;
  city?: string | null;
  plan: string;
  rankLevel: number;
  rankPoints: number;
  tasksCompleted: number;
  totalEarned: number;
  createdAt: Date;
}